+++ src/app/(autenticated)/(main)/page.tsx
@@ -1,3 +1,4 @@
+import AudioRecord from '@/components/record/AudioRecord';
 import InstallPrompt from '@/components/shared/InstallPrompt';
 import { authOptions } from '@/utils/authOptions';
 import { Metadata } from 'next';
@@ -16,8 +17,9 @@ export default async function Page() {
   }
 
   return (
-    <div className='flex justify-center items-center size-full'>
+    <div className='flex flex-col gap-4 justify-center items-center size-full'>
       <InstallPrompt />
+      <AudioRecord session={session} />
     </div>
   );
 }
