+++ src/app/(autenticated)/diary/page.tsx
@@ -1,8 +1,7 @@
+import DiaryList from '@/components/diary/DiaryList';
+import { authOptions } from '@/utils/authOptions';
 import { getServerSession } from 'next-auth';
 import { redirect } from 'next/navigation';
-import Link from 'next/link';
-import { authOptions } from '@/utils/authOptions';
-import AudioRecord from '@/components/write/AudioRecord';
 
 export default async function page() {
   const session = await getServerSession(authOptions);
@@ -11,11 +10,5 @@ export default async function page() {
     redirect('/auth/signin');
   }
 
-  return (
-    <main>
-      <h1>일기 녹음 페이지</h1>
-      <AudioRecord session={session} />
-      <Link href='/'>홈으로</Link>
-    </main>
-  );
+  return <DiaryList session={session} />;
 }
