+++ src/app/actions.ts
@@ -173,3 +173,21 @@ export async function getDiaries(userId: string) {
 
   return data;
 }
+
+// diary 테이블에서 오늘 날짜의 데이터를 가져오는 함수
+export async function getDiaryAtDate(userId: string, createdAt: string) {
+  const supabase = await createClient();
+  const { data, error } = await supabase
+    .from('test_diary')
+    .select()
+    .eq('user_id', userId)
+    .gte('created_at', `${createdAt}T00:00:00.000Z`)
+    .lte('created_at', `${createdAt}T23:59:59.999Z`);
+
+  if (error) {
+    console.error(error);
+    throw error;
+  }
+
+  return data;
+}
