+++ src/components/record/AudioRecord.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import RecordButton from './RecordButton';
+import { Session } from 'next-auth';
+
+export default function AudioRecord({ session }: { session: Session }) {
+  return (
+    <div>
+      <RecordButton session={session} />
+    </div>
+  );
+}
